pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[project]
classifiers = [
    'Development Status :: 5 - Production/Stable',
    'Intended Audience :: Science/Research',
    'Topic :: Scientific/Engineering :: Medical Science Apps.',
    'Natural Language :: English',
]
dependencies = [
    'pyyaml>=6.0.1',
    'numpy>=1.18.0',
    'tqdm>=4.45.0',
    'click>=8.1.7',
    'xnippet', #'xnippet @ git+https://github.com/dvm-shlee/xnippet.git',
    'reshipe', #'reshipe @ git+https://github.com/dvm-shlee/reshipe.git'
]
description = "Bruker PvDataset Loader"
license = {text = "GNLv3"}
dynamic = ["version"]
maintainers = [
    {name = "SungHo Lee", email = 'shlee@unc.edu'}
]
name = "brkraw"
readme = "README.md"
requires-python = ">=3.7"
keywords = [
    'bruker', 
    'data_handler', 
    'converter', 
    'administrator_tool', 
    'extensible', 
]

[project.urls]
Homepage = "https://brkraw.github.io"

[tool.hatch.metadata]
allow-direct-references = true

[project.optional-dependencies]
all = ["brkraw[tonii]", "brkraw[viewer]"]

tonii = [
    'nibabel>=3.0.2',
]

viewer = [
    'pillow>=7.1.1'
]

dev = [
    "flake8",
    "pytest",
    "types-PyYAML",
]

[tool.hatch.version]
path = "brkraw/__init__.py"
style = "pep440"

[project.scripts]
brkraw = "brkraw.scripts:main"
brkraw-tonii = "brkraw.scripts.tonifti:tonii"
brkraw-manager = "brkraw.scripts.manager:manager"
brkraw-viewer = "brkraw.scripts.viewer:viewer"
brkraw-bids = "brkraw.scripts.bids:bids"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "DEBUG"
